+++ ably/ably_test.go
@@ -1,26 +1,12 @@
 package ably_test
 
 import (
-	"errors"
 	"fmt"
 	"io"
-	"os"
 	"testing"
-	"time"
 
 	"github.com/ably/ably-go/ably"
 )
-
-var timeout = 5 * time.Second
-
-func init() {
-	ably.Log.Level = ably.LogVerbose
-	if s := os.Getenv("ABLY_TIMEOUT"); s != "" {
-		if t, err := time.ParseDuration(s); err == nil {
-			timeout = t
-		}
-	}
-}
 
 func nonil(err ...error) error {
 	for _, err := range err {
@@ -29,22 +15,6 @@ func nonil(err ...error) error {
 		}
 	}
 	return nil
-}
-
-func wait(res ably.Result, err error) error {
-	if err != nil {
-		return err
-	}
-	errch := make(chan error)
-	go func() {
-		errch <- res.Wait()
-	}()
-	select {
-	case err := <-errch:
-		return err
-	case <-time.After(timeout):
-		return errors.New("waiting on Result timed out after " + timeout.String())
-	}
 }
 
 func safeclose(t *testing.T, closers ...io.Closer) {
