+++ ably/export_test.go
@@ -1,0 +1,46 @@
+package ably
+
+import "net/http"
+
+func (p *PaginatedResult) BuildPath(base, rel string) (string, error) {
+	return p.buildPath(base, rel)
+}
+
+func (opts *ClientOptions) RestURL() string {
+	return opts.restURL()
+}
+
+func (opts *ClientOptions) RealtimeURL() string {
+	return opts.realtimeURL()
+}
+
+func (c *RestClient) Post(path string, in, out interface{}) (*http.Response, error) {
+	return c.post(path, in, out)
+}
+
+func DecodeResp(resp *http.Response, out interface{}) error {
+	return decodeResp(resp, out)
+}
+
+func ErrorCode(err error) int {
+	return code(err)
+}
+
+// MustRealtimeClient is like NewRealtimeClient, but panics on error.
+func MustRealtimeClient(opts *ClientOptions) *RealtimeClient {
+	client, err := NewRealtimeClient(opts)
+	if err != nil {
+		panic("ably.NewRealtimeClient failed: " + err.Error())
+	}
+	return client
+}
+
+// GetAndAttach is a helper method, which returns attached channel or panics if
+// the attaching failed.
+func (ch *Channels) GetAndAttach(name string) *RealtimeChannel {
+	channel := ch.Get(name)
+	if err := wait(channel.Attach()); err != nil {
+		panic(`attach to "` + name + `" failed: ` + err.Error())
+	}
+	return channel
+}
